Moved Averages to nav
