+++ client/components/nav/Averages.jsx
