+++ client/components/nav/AveragesStars.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import Stars from './Stars';
+import Stars from '../Stars';
 
 const AveragesStars = (props) => {
   const classes = { ...props };
   return (
     <span>
       <span>Overall</span>
-      <Stars average={classes.average === NaN ? 0 : classes.average} />
+      <Stars average={Number.isNaN(classes.average) ? 0 : classes.average} />
       <span>{classes.average}</span>
     </span>
   );
